+++ asto-core/src/main/java/com/artipie/asto/LoggingStorage.java
@@ -35,6 +35,11 @@ public final class LoggingStorage implements Storage {
     private final Storage storage;
 
     /**
+     * Storage string identifier.
+     */
+    private final String id;
+
+    /**
      * Ctor.
      *
      * @param storage Delegate storage.
@@ -55,6 +60,7 @@ public final class LoggingStorage implements Storage {
         this.logger = Logger.getLogger(
             this.storage.getClass().getCanonicalName()
         );
+        this.id = String.format("Logging: %s", this.storage.identifier());
     }
 
     @Override
@@ -160,6 +166,11 @@ public final class LoggingStorage implements Storage {
                 return result;
             }
         );
+    }
+
+    @Override
+    public String identifier() {
+        return this.id;
     }
 
     /**